new changes

+++ src/actions/domOperation.js
@@ -6,7 +6,6 @@ const url = "https://todo-backend-server-v1.herokuapp.com/tasks";
 
 export const getAllTasks = async () => {
     let allTasks = await taskapiCall(`${url}`);
-    console.log(allTasks);
     allTasks.forEach((item) => {
         createTask(item);
        
@@ -16,8 +15,13 @@ export const getAllTasks = async () => {
 
 export const addTasks = async (event) => {
     event.preventDefault();
+    if (!navigator.onLine) {
+        alert("you are offline!");
+        return;
+    }
     let taskField = document.getElementById("taskField");
     let inputValue = taskField.value;
+    console.log(inputValue);
     taskField.value = "";
     let taskObj = {
         method: 'POST',
@@ -29,11 +33,14 @@ export const addTasks = async (event) => {
         }
     }
     const addTask = await taskapiCall(`${url}`, taskObj);
-    console.log(addTask);
     createTask(addTask);
 }
 
 export const deleteTasks = (e) => {
+    if (!navigator.onLine) {
+        alert("you are offline!");
+        return;
+    }
     let taskObj = {
         method: 'DELETE'
     }
@@ -45,6 +52,10 @@ export const deleteTasks = (e) => {
 }
 
 export const updateTasks = (e) => {
+    if (!navigator.onLine) {
+        alert("you are offline!");
+        return;
+    }
     const enableInput = e.target.parentElement.childNodes[1];
     enableInput.disabled = false;
 
@@ -52,6 +63,10 @@ export const updateTasks = (e) => {
 }
 
 export const doneEditTasks = async (e) => {
+    if (!navigator.onLine) {
+        alert("you are offline!");
+        return;
+    }
     const enableInput = e.target.parentElement.childNodes[1];
     const parentelem = e.target.parentElement;
     const id = e.target.parentElement.id;
@@ -78,18 +93,24 @@ export const doneEditTasks = async (e) => {
 
 
 export const taskCompleted = async (e) => {
+    if (!navigator.onLine) {
+        alert("you are offline!");
+        return;
+    }
     const enableInput = e.target.parentElement.childNodes[1];
+    const complete = e.target.parentElement.childNodes[1];
     const editDisabled = e.target.parentElement.childNodes[3];
+    const deleteBtn = e.target.parentElement.childNodes[2];
     const doneDisabled = e.target.parentElement.childNodes[4];
     const id = e.target.parentElement.id;
     editDisabled.disabled = true;
     doneDisabled.disabled = true;
     enableInput.style.textDecoration = "line-through";
     e.target.parentElement.style.backgroundColor = "#ae5f75";
-    delBtn.style.color="white";
-    editBtn.style.color="white";
-    doneEdit.style.color="white";
-    completeBtn.style.color="white";
+    deleteBtn.style.color="white";
+    editDisabled.style.display="none";
+    doneDisabled.style.color="white";
+    complete.style.color="white";
 
     console.log(editDisabled);
     let inputValue = enableInput.value;
